+++ libs/core-ui/src/index.ts
@@ -33,6 +33,7 @@ export * from "./lib/util/getCohortFilterCount";
 export * from "./lib/util/getTreatmentBarChartOptions";
 export * from "./lib/util/getDependencyChartOptions";
 export * from "./lib/util/IGenericChartProps";
+export * from "./lib/util/ImageStatisticsUtils";
 export * from "./lib/util/initializeOfficeFabric";
 export * from "./lib/util/ModelExplanationUtils";
 export * from "./lib/util/Never";
